backend/requirements.txt
Django==5.0.2
mysqlclient==2.2.4
pymysql>=1.0.2
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-environ==0.11.2
Pillow==10.2.0
django-widget-tweaks==1.4.12
django-allauth==0.61.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
whitenoise==6.6.0
django-debug-toolbar==4.3.0
django-extensions==4.1
python-decouple==3.8

